feat: move lobby script into 1 directory
