+++ Assets/Scenes/StartAndLobbyAndWaiting.unity
@@ -4953,7 +4953,7 @@ GameObject:
   - component: {fileID: 1439638006}
   - component: {fileID: 1439638005}
   m_Layer: 5
-  m_Name: CreateButtontton
+  m_Name: CreateButton
   m_TagString: Untagged
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
@@ -6673,7 +6673,7 @@ GameObject:
   - component: {fileID: 1839378744}
   - component: {fileID: 1839378743}
   m_Layer: 5
-  m_Name: CancelButtontton
+  m_Name: CancelButton
   m_TagString: Untagged
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
